+++ src/main/java/com/telepathicgrunt/ultraamplifieddimension/modInit/UADProcessors.java
@@ -1,0 +1,28 @@
+package com.telepathicgrunt.ultraamplifieddimension.modInit;
+
+import com.telepathicgrunt.ultraamplifieddimension.UltraAmplifiedDimension;
+import com.telepathicgrunt.ultraamplifieddimension.world.processors.*;
+import net.minecraft.structure.processor.StructureProcessorType;
+import net.minecraft.util.Identifier;
+import net.minecraft.util.registry.Registry;
+
+public class UADProcessors {
+
+    public static StructureProcessorType<WaterloggingFixProcessor> WATER_FIX_PROCESSOR = () -> WaterloggingFixProcessor.CODEC;
+    public static StructureProcessorType<ReplaceAirOnlyProcessor> REPLACE_AIR_ONLY_PROCESSOR = () -> ReplaceAirOnlyProcessor.CODEC;
+    public static StructureProcessorType<ReplaceLiquidOnlyProcessor> REPLACE_LIQUIDS_ONLY_PROCESSOR = () -> ReplaceLiquidOnlyProcessor.CODEC;
+    public static StructureProcessorType<RemoveFloatingBlocksProcessor> REMOVE_FLOATING_BLOCKS_PROCESSOR = () -> RemoveFloatingBlocksProcessor.CODEC;
+
+    public static StructureProcessorType<WallVinePostProcessor> WALL_VINE_POST_PROCESSOR = () -> WallVinePostProcessor.CODEC;
+    public static StructureProcessorType<CeilingVinePostProcessor> CEILING_VINE_POST_PROCESSOR = () -> CeilingVinePostProcessor.CODEC;
+
+    public static void registerProcessors() {
+        Registry.register(Registry.STRUCTURE_PROCESSOR, new Identifier(UltraAmplifiedDimension.MODID, "water_fix_processor"), WATER_FIX_PROCESSOR);
+        Registry.register(Registry.STRUCTURE_PROCESSOR, new Identifier(UltraAmplifiedDimension.MODID, "replace_air_only_processor"), REPLACE_AIR_ONLY_PROCESSOR);
+        Registry.register(Registry.STRUCTURE_PROCESSOR, new Identifier(UltraAmplifiedDimension.MODID, "replace_liquids_only_processor"), REPLACE_LIQUIDS_ONLY_PROCESSOR);
+        Registry.register(Registry.STRUCTURE_PROCESSOR, new Identifier(UltraAmplifiedDimension.MODID, "remove_floating_blocks_processor"), REMOVE_FLOATING_BLOCKS_PROCESSOR);
+
+        Registry.register(Registry.STRUCTURE_PROCESSOR, new Identifier(UltraAmplifiedDimension.MODID, "wall_vine_post_processor"), WALL_VINE_POST_PROCESSOR);
+        Registry.register(Registry.STRUCTURE_PROCESSOR, new Identifier(UltraAmplifiedDimension.MODID, "ceiling_vine_post_processor"), CEILING_VINE_POST_PROCESSOR);
+    }
+}